Refactor: QRCodeSheetView 생성 및 관련 코드 삭제
