+++ 2024-MacC-M4-6princess/2024-MacC-M4-6princess/Source/View/ImageOutput/IOView.swift
@@ -17,15 +17,7 @@ struct IOView: View {
     
     @Environment(\.presentationMode) var presentationMode: Binding<PresentationMode>
     @StateObject var viewModel = IOViewModel()
-    @State var isAnimating = false
-    @State var isSave = false
-    @State private var isUploading = false
-    
-    @State private var qrCodeImage: UIImage? = nil
-    @State private var uploadError: String? = nil
-    @State var showQRSheet = false
-    @State private var uploadedImagePath: String? = nil // 업로드된 이미지의 경로
-    
+
     var body: some View {
         GeometryReader { geometry in
             VStack{
@@ -53,14 +45,9 @@ struct IOView: View {
                 // 후보정 레이어 편집 뷰
                 canvasView
                     .onAppear{
-                        isAnimating = true
                         viewModel.saveRenderedView(content: canvasView, motionManager: motionManager)
                         viewModel.saveAnimate = true
-                        DispatchQueue.main.asyncAfter(deadline: .now() + 1) {
-                            isSave = true
-                        }
                         print("canvasView onAppear")
-                        uploadImage()
                     }
                     .applyIf(motionManager.currentOrientation != .portrait && motionManager.currentOrientation != .portraitUpsideDown) { original in
                         original.modifier(
@@ -142,87 +129,18 @@ struct IOView: View {
             
             print("body onAppear")
         }
-        //        .onDisappear {
-        //            // QR 보기를 안눌렀으면 파이어베이스에서 사진 삭제
-        //                   deleteUploadedImageIfNeeded()
-        //               }
         .alert(isPresented: $viewModel.showAlert) {
             Alert(title: Text("오류 발생"), message: Text(viewModel.alertMessage), dismissButton: .default(Text("확인")))
         }
         
         // 상단 툴바
         .navigationBarBackButtonHidden()
-        .sheet(isPresented: $showQRSheet) {
-            QRCodeSheetView(qrCodeImage: qrCodeImage)
-        }
     }
     var canvasView: some View {
         IOCanvasView(viewModel: viewModel)
     }
-    private func uploadImage() {
-        guard let selectedImage = viewModel.compositeImage else { return }
-        isUploading = true
-        StorageManager.shared.uploadImage(image: selectedImage) { result in
-            DispatchQueue.main.async {
-                isUploading = false
-                switch result {
-                case .success(let url):
-                    self.qrCodeImage = QRCodeGenerator.shared.generateQRCode(from: url.absoluteString)
-                    self.uploadedImagePath = url.path // 업로드된 이미지 경로 저장
-                case .failure(let error):
-                    self.uploadError = error.localizedDescription
-                }
-            }
-        }
-    }
-    private func deleteUploadedImageIfNeeded() {
-        guard let path = uploadedImagePath else { return }
-        StorageManager.shared.deleteImage(path: path) { result in
-            switch result {
-            case .success:
-                print("Image successfully deleted from Firebase.")
-            case .failure(let error):
-                print("Failed to delete image: \(error.localizedDescription)")
-            }
-        }
-    }
 }
 
 
 
-import SwiftUI
-import FirebaseAnalytics
 
-struct QRCodeSheetView: View {
-    let qrCodeImage: UIImage?
-    @Environment(\.dismiss) var dismiss
-    
-    var body: some View {
-        VStack {
-            if let qrCodeImage = qrCodeImage {
-                // QR 코드 이미지 표시
-                Image(uiImage: qrCodeImage)
-                    .resizable()
-                    .scaledToFit()
-                    .frame(width: 250, height: 250)
-                    .padding()
-            } else {
-                // QR 코드 로딩 중일 때 ProgressView 표시
-                ProgressView("QR 코드를 생성 중입니다...")
-                    .padding()
-            }
-            
-            Button("닫기") {
-                dismiss()
-            }
-            .padding()
-            .foregroundColor(.white)
-            .background(Color.pointPink)
-            .cornerRadius(8)
-        }
-        .padding()
-        .onAppear {
-            Analytics.logEvent("A99_QR보기", parameters: nil) // 라벨링 추가
-        }
-    }
-}
